[REFACTOR] : EditProfileContainer.tsx를 Page.tsx로 통합
